# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "API documentation for Blog",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/Blog/public",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/users": {
            "get": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Users and all available roles.",
                "description": "Show all Users with their user roles and permissions, and all available roles.",
                "operationId": "eaedf6aa84fde38a08bf5d87308169f7",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "All Users and Their Roles and All Available Roles"
                                        },
                                        "data": {
                                            "properties": {
                                                "users": {
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/UserResource"
                                                            }
                                                        },
                                                        "links": {
                                                            "properties": {
                                                                "first": {
                                                                    "type": "string",
                                                                    "example": "http://example.com/api/posts?page=1"
                                                                },
                                                                "last": {
                                                                    "type": "string",
                                                                    "example": "http://example.com/api/posts?page=5"
                                                                },
                                                                "prev": {
                                                                    "type": "string",
                                                                    "example": null,
                                                                    "nullable": true
                                                                },
                                                                "next": {
                                                                    "type": "string",
                                                                    "example": "http://example.com/api/posts?page=2",
                                                                    "nullable": true
                                                                }
                                                            },
                                                            "type": "object"
                                                        },
                                                        "meta": {
                                                            "properties": {
                                                                "current_page": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "from": {
                                                                    "type": "integer",
                                                                    "example": 1
                                                                },
                                                                "last_page": {
                                                                    "type": "integer",
                                                                    "example": 5
                                                                },
                                                                "path": {
                                                                    "type": "string",
                                                                    "example": "http://example.com/api/users"
                                                                },
                                                                "links": {
                                                                    "type": "array",
                                                                    "items": {
                                                                        "properties": {
                                                                            "url": {
                                                                                "type": "string",
                                                                                "example": "http://localhost/Blog/public/api/users?page=1",
                                                                                "nullable": true
                                                                            },
                                                                            "label": {
                                                                                "type": "string",
                                                                                "example": "1"
                                                                            },
                                                                            "active": {
                                                                                "type": "boolean",
                                                                                "example": true
                                                                            }
                                                                        },
                                                                        "type": "object"
                                                                    }
                                                                },
                                                                "per_page": {
                                                                    "type": "integer",
                                                                    "example": 20
                                                                },
                                                                "to": {
                                                                    "type": "integer",
                                                                    "example": 20
                                                                },
                                                                "total": {
                                                                    "type": "integer",
                                                                    "example": 50
                                                                }
                                                            },
                                                            "type": "object"
                                                        }
                                                    },
                                                    "type": "object"
                                                },
                                                "roles": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/RoleResource"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized Action"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/change-role/{user}": {
            "put": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Change User roles",
                "description": "Change User roles.",
                "operationId": "b1cbe8a3a38c2409fbca261276ca31fb",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "role_ids"
                                ],
                                "properties": {
                                    "role_ids": {
                                        "description": "exists in table roles,id",
                                        "type": "array",
                                        "items": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Roles changed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Roles Changed Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "User is deactivated or Unauthorized Action",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Error"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "The change role is Prevented Because the user account is deactivated, \n     *                 or you do not have the permission to change User Role or change Role very important User."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "User not found"
                    },
                    "422": {
                        "description": "Validation Error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/ban/{id}": {
            "delete": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Baning User",
                "description": "Baning specified User.",
                "operationId": "e87799a810c122faa49d80a06e846890",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User is Banned",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User is Banned Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized Action"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/activate/{id}": {
            "get": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Activating User",
                "description": "Activating specified User.",
                "operationId": "ac782a73262fc54e41818ed3526fdf3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User Activated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User Activated Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized Action"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/delete/{id}": {
            "delete": {
                "tags": [
                    "Authorization"
                ],
                "summary": "Delete User",
                "description": "Delete specified User.",
                "operationId": "36d92c1c6aecf32e1b6aed40b115aa7a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User is Deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "Success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User is Deleted Successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Unauthorized Action"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "RoleResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "role name"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "users": {
                        "description": "Loaded only if relation users is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/UserResource"
                        }
                    },
                    "permissions": {
                        "description": "Loaded only if relation permissions is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PermissionResource"
                        }
                    }
                },
                "type": "object"
            },
            "UserResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "user name"
                    },
                    "email": {
                        "type": "string",
                        "example": "name@test.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "user phone"
                    },
                    "lang": {
                        "type": "string",
                        "example": "user lang"
                    },
                    "favicon": {
                        "type": "string",
                        "example": "user favicon"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "active": {
                        "type": "boolean",
                        "example": true
                    },
                    "posts": {
                        "description": "Loaded only if relation posts is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PostResource"
                        }
                    },
                    "comments": {
                        "description": "Loaded only if relation comments is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CommentResource"
                        }
                    },
                    "roles": {
                        "description": "Loaded only if relation roles is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/RoleResource"
                        }
                    },
                    "permissions": {
                        "description": "Loaded only if relation roles is loaded",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/PermissionResource"
                        }
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}